requirements.txt
flask==3.0.3
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
gunicorn==22.0.0
